Http.with_headers, SSL, and automatic JSON parsing, oh my!
I really do like to cowboy crap together sometimes. The first example I picked
for trying to send a JSON accept header with was an HTTPS github URL. https
is totally broken with Net::HTTP, as in if you give it an https URL, it will
try to make the request without SSL unless you explicitly tell it to use SSL.
THIS WILL NOT STAND.

The whole point of this commit was to add with_headers, but I threw all the
other stuff that was annoying me with Net::HTTP in so the tests would pass.
Mea culpa.