Http::Response handling fixups
Tyler's changes add a Http::Response object and round trip stuff done
using the Http module through that. But Http is really designed to be
a shortcut for the common case of just wanting to grab resources off
the web.

Head requests are a good case for an exception to that behavior (as are
other requests, potentially). I'd like to examine each verb one by one
and pick the appropriate value to return from each one.

This change alters the way Http::Response objects are built to be more
like a Struct. Eventually I'd like to have the parser build these objects,
but for now Http is manually reconstructing them from Net::HTTP's response
objects.