Remove helper methods from the Client itself (SRP)
The Client class already handles Net::HTTP bridging and most of
the semantics of the options. It makes sense to frees it from also
having to provide http helpers, which are available through Chainable.

Removing the helper methods does not hurt (provided Client is part of
the private API?). Client#request become *the method* for which injection
will take place later.