Properly add the :response => :auto case.
Previously Chainable implemented a different heuristics than Client
for :response. Chainable used :object on HEAD and :parsed_body on
other verbs. Client always used :object.

This commit removes this difference and introduces :auto as option
value. It is the default one, it is implemented in Client directly
and it corresponds to the Chainable behavior above. This means that
{:response => :object} is necessary on Client to get its previous
behavior (broken private API, therefore).

By the way, we could restore the previous behavior simply by setting
:object as default in Options and forcing something else through
default_options in Chainable and/or Http. It seemed simpler to get the
auto behavior immediately in Client, though.