Drop YARDStick threshold to 58.8
Drop in coverage is legit. Actually it became higher, but due to bug in YARD
itself (see https://github.com/lsegal/yard/issues/726), seems like macros with
doc reference treated as "no docs" (as it really generates empty description):

    @!method foo
      (see Other#foo)

The above should copy all description (params and return and such) from
`Other#foo` to current scope's `#foo`, and its not at the moment. Still,
this YARD syntax is legit and should remain as is. Waiting for that bug
to be resolved. :D