Unified exceptional/exceptionless non-blocking I/O
The main motivation for this change is because JRuby 9000 supports exceptionless
non-blocking I/O for TCPSockets but not for SSLSockets. Both accept the
":exception => false" syntax, so it's harmless to pass it to SSLSockets even
though they silently ignore it.

It also adds handling non-blocking reads during writes and writes during reads,
which is necessary for SSLSockets to work correctly on all platforms (i.e. an
SSL handshake might require performing a read for a write to complete, or vice
versa)

I'm pleased with the result in as much as it provides a single approach which
covers all cases. The readpartial / write and wait_readable_or_timeout /
wait_writable_or_timeout methods are almost identical at this point and are
clearly ripe for some DRY-out refactoring.