Use io/wait on platforms where it's available
IO.select is an expensive call, because it has to build an fd_set (large
bitfield) from an array each time we invoke it.

We're just interested in IO readiness, so on Ruby 2.0+ we can use the io/wait
API instead, which adds wait_readable/wait_writable to IO objects.