Use object composition rather than inheritance for HTTP::URI
The Addressable gem has been the source of a number of bugs and
performance problems. Perhaps in the future it might be worth
avoiding using it where ::URI would suffice, or getting rid of
it entirely, but as things stand it's the parent class for
HTTP::URI, which prevents us from doing that without making a
breaking release of the library.

This replaces inheritence with object composition: Forwardable
is used to delegate an explicit set of methods to Addressable::URI.
This should make it easier to work around Addressable bugs or
potentially even use ::URI as the internal representation when
URIs are simple enough.