Avoid a dependency cycle between Client and Connection classes.
I found out that sometimes you can get this error when using the gem:

    from /.../gems/http-2.0.1/lib/http.rb:8:in  `<top (required)>'
    from /.../gems/http-2.0.1/lib/http.rb:8:in  `require'
    from /.../gems/http-2.0.1/lib/http/client.rb:7:in  `<top (required)>'
    from /.../gems/http-2.0.1/lib/http/client.rb:7:in  `require'
    from /.../gems/http-2.0.1/lib/http/connection.rb:4:in  `<top (required)>'
    from /.../gems/http-2.0.1/lib/http/connection.rb:4:in  `require'

The problem was that 'http.rb' was requiring 'http/client.rb' that at
the same time was requiring 'http/connection.rb' which was requiring
'http/client.rb' again.

I think that in general is not very good to have a cycle like this in
the dependency graph of the code, so I decided to remove that dependency
by moving the constants KEEP_ALIVE and CLOSE to the Connection class.
This way Connection is now independent from Client.

Maybe this constants could go to another module but I couldn't think in
a better place.