Enable request body streaming with an IO object
This change adds the ability to stream content of an IO object (an
object that responds to #read) into the request body. This is convenient
when you want to send large amounts of data as the request body, but you
don't want to load all of it into memory.

  HTTP.post("http://example.org/upload", body: File.open("video.mp4"))
  HTTP.post("http://example.org/upload", body: StringIO.new("file data"))

It was already possible to send the request body in chunks by passing an
Enumeable to :body, but that's only defined for "Transfer-Encoding:
chunked" requests.

This will also allow us to extend form_data.rb to stream the
multipart-encoded request, instead having to load the whole body into
memory before writing it to the socket.