Allow chunked requests for any body type
Before creating chunked requests was only possible with Enumerable body.
If "Transfer-Encoding: chunked" was set but the body wasn't Enumerable,
request would be calculated as if it was a regular request, which meant
inclusion of "Content-Length" header and writing request body as is. If
body was an Enumerable but "Transfer-Encoding: chunked" wasn't set, an
error would be raised.

This commit enables sending chunked requests with any body type, by
abstracting how the request body is read into an #each method. If
"Transfer-Encoding: chunked" is detected, the yielded chunks are encoded
accordingly before writing, otherwise the chunks are written as is.

This allows making a streaming chunked request using an IO object, where
file would be written in chunks of 16 KB, which wasn't previously
possible.