Let Host be automatically inferred on redirects
When redirecting on "http://localhost:<port>" URLs, the Host will be
correctly set to "http://localhost:<port>" on the initial request, but
on following the redirect the port is lost, because we're only assigning
the host.

Therefore when we initialize the follow-up request object we de-assign
"Host", and let HTTP::Request set it in the same way as it is set on the
initial request.