Make auto deflate work with any request body
Previously auto deflating worked only for string bodies, so we
generalize the implementation to work with any request body, by relying
on the Request::Body abstraction.

We also improve memory usage by yielding deflated chunks instead of
returning the whole deflated body.

On "Transfer-Encoding: chunked" requests the compressed body will be
directly streamed to the socket as it is deflated, while on regular
requests all of the body has to be deflated first so that we can obtain
the body size for the "Content-Length" header before we start writing
content.