Return response body chunks in specified encoding
MRI will return content read from the socket in the ASCII-8BIT (binary)
encoding, whereas JRuby will return it in the UTF-8 encoding. In
whichever encoding the body is retrieved, we want to force its encoding
to the one specified (charset response header if present, otherwise
binary). This is already the behaviour in Response#to_s, we just extend
it to Response#readpartial as well.

Fixes #413