Use #readpartial for reading content if available
IO#readpartial is preferable in cases where the IO is a stream such as a
pipe or a socket. IO#readpartial differs from IO#read in situations
where there is still some data immediately available, but not the number
of bytes requested. IO#read would in this case block until the requested
number of bytes is available, whereas IO#readpartial would immediately
return the available data, and only block if there is no data available.

Another technical difference is that, while IO#read simply returns nil
when it has reached EOF, IO#readpartial raises an EOFError, so we have
to rescue it.

With IO#readpartial uploading from pipes should now be faster, because
any data that is immediately available will be written to the TCP
socket, so the upload is less blocked by the pipe retrieving data. This
is also useful with the Down gem, which provides an IO object around a
remote file, because again uploading is less blocked by the remote file
being downloaded.

This IO#readpartial preference is even baked into Ruby itself;
`IO.copy_stream` will also prefer IO#readpartial over IO#read.