Use a buffer string into which chunks will be read
IO#read and IO#readpartial accept an optional buffer string as a second
argument. This will make it so that, instead of each chunk being
allocated as a new string, each chunk is read into the buffer string
(replacing any previous value). This means that no new strings will be
allocated other than the buffer string on upload, so it's much more
memory-efficient and gives less work to the GC.

Previously we couldn't use a buffer string because of how JRuby
implements Enumerators (it fetches the next chunk eagerly, replacing the
one being returned). That's why we changed the implementation not to use
Enumerators.