Use IO.copy_stream for yielding chunks
IO.copy_stream already has the optimized behaviour of using #readpartial
when available and a buffer string. So we create a BlockIO object that
responds to #write, calling the underlying block with the given data,
and returns the amount of written bytes. This BlockIO object can then be
used as the "destination IO" in IO.copy_stream.