Remove Enumerator from Request::Body#each
Using Enumerator interface like Enumerator#next wouldn't work correctly
on JRuby with IOs, because JRuby eagerly loads the next element before
it returns the current, and because each next chunk overwrites the
previous one, it is impossible to retrieve first chunk via
Enumerable#next.

https://github.com/jruby/jruby/issues/4583

In order to avoid potential bugs in the future of someone wanting to use
this Enumerator interface, we remove this entirely and mandate people to
pass in a block.