Raise ConnectionError when writing to socket fails
Writing to the TCP socket can also fail with various exceptions, just
like when opening or reading from it. For example, it happened to me
that an Errno::EPIPE exceptions was raised during file upload,
presumably because the server closed the connection early.

In order to be consistent with re-raising low-level exceptions as
HTTP::ConnectionError when opening or reading from the TCP socket,
we also add the same behaviour when writing to the socket.