Reduce memory usage when reading response body
Currently for each chunk read from the socket a new string is being
allocated. This is not necessary, because we're feeding that string
object into the HTTP parser and never using it again.

So we initialize a buffer string and pass it in when reading chunks.
This will make each chunk be read into the same buffer object, no new
string will be allocated in that operation.

In my benchmark this halves the allocated memory when downloading
large response bodies.