Refactor AutoInflate/Deflate into generic features
This makes HTTP::Client, Request and Response unaware of the
implementation details behind inflating and deflating stream bodies.
Instead, the client iterates over all enabled features on request and
response, calling `#wrap_request` or `#wrap_response`, which is expected
to return the Request or Response as-is if nothing should be done, or
returns a new Request or Response object after having done whatever
transformation the feature requires.