Response#uri can be set directly, or pulled from the Request
During normal operation, the Response#uri will be pulled from
the Request passed into the Response. However, we still allow explicitly
setting a `:uri` option when creating a Response, for backwards compatibility.

This was motivated by rubocop/code climate complaining that `Client#perform`
was now 1-line too long. I'll defer to the project maintainers to decide
if that is a good thing.