RuboCop upgrade (#548)
* Update Rubocop and run auto-fixes

In an attempt to do some house keeping this pushed the rubocop version
up by ten and fixes everything that is correctable from an auto-fix
standpoint. These changes should be safe if we assume that rubocop's
auto-fix changes are safe.

* Disable cop around black hole

This black hole method should not call up to super. That is the purpose
of a black hole object.

* Do not take certain changes to maintain backwards compatibility

Do not take the Unfreeze string cop because it is only supported in Ruby
2.3 and above.

* Fix class_eval to be friendlier

This came up in Style::EvalWithLocation. It makes backtraces easier to
follow.

* Selectively disabled cops for methods and classes

Rubocop now requires that cops are scoped tighter. This reformats it so
that only the specific method or class has the cop disabled.

* Enforce implicit on Style/RescueStandardError

Per ixti's code review, the maintainers prefer this enforced style so I
am changing to the requested style.

* Simplify code to ternary

This code could easily be collapsed to a ternary per ixti's request so
that is what I am doing.