Bump to RuboCop latest version and run autocorrect
This is the latest version of rubocop with most of the auto-corrects
run. There are a few notes on this:
- Performance cops have been refactored out so they require being
installed in a new gem. That gem has been included so that we can keep
performance cops.
- I left out the Alight Hash cop because the code base has it going both
ways. I want to put that in its own commit so that it is easy to change
or reverse if the maintainers want to go one way or another.