Sketch of what max_repeats could look like.
This was inspired by an improvement requested in #544.

Basically we want there to be some configurability around how long
things are allowed to loop for before we stop or blow up.

The way the code proposed currectly works is that if max_repeats is set
it will return the latest redirect instead of blowing up. However once I
wrote it I wondered if we should instead blow up.

I'm opening a draft PR to get some feedback on how this should behave
and once we figure out what we want to do I can go back and make this
more robust and write specs for it.