Duplicate options Hash in .timeout call
When `Hash` is passed to HTTP.timeout, it's been mutated. So when you pass
a frozen `Hash` it raises an exception.

Here is an example of the code:
```
timeout_settings = { connect: 5, write: 2, read: 10 }.freeze
HTTP.timeout(timeout_settings).get('http://example.com')
```

In this change, I have added a call to .dup on options Hash to create a
copy and avoid mutation of original Hash.