Recognize cookies set by redirect
When a redirect response arrives, the browser will recognize
`Set-Cookie` headers the same as any other request/response. This PR
enables the Redirector to correctly follow the cookie hand-offs.

I encountered this situation out in the wild. This PR, as-is, seems to
fix the issue. I haven't bothered to find the relevant spec.